Reorganize tests (#169)
