--- test/mocha.opts
@@ -1,3 +1,0 @@
---require dotenv/config
---require ts-node/register
---require source-map-support/register
