+++ test/projects/hardhat-merge-project/hardhat.config.ts
@@ -1,8 +1,7 @@
 // We load the plugin here.
 import { HardhatUserConfig } from "hardhat/types";
-import "@nomiclabs/hardhat-waffle";
 
-import "../../src/index";
+import "../../../src/index";
 
 const config: HardhatUserConfig = {
   solidity: "0.5.8",
