+++ test/projects/hardhat-truffle-project/contracts/DuplicateA.sol
