+++ test/projects/hardhat-truffle-project/contracts/DuplicateB.sol
