+++ test/projects/hardhat-truffle-project/contracts/MetaCoin.sol
