+++ test/projects/hardhat-truffle-project/contracts/VariableConstructor.sol
