+++ test/projects/hardhat-truffle-project/contracts/VariableCosts.sol
