+++ test/projects/hardhat-truffle-project/test/metacoin.js
