+++ test/tasks/merge.ts
@@ -1,14 +1,14 @@
 import fs from "fs"
 import path from "path"
-import { TASK_GAS_REPORTER_MERGE } from "../src/task-names";
+import { TASK_GAS_REPORTER_MERGE } from "../../src/task-names";
 // tslint:disable-next-line no-implicit-dependencies
 import { assert } from "chai";
-import { useEnvironment } from "./helpers";
+import { useEnvironment } from "./../helpers";
 
 const loadJsonFile = (filepath) => JSON.parse(fs.readFileSync(filepath, "utf-8"))
 
 describe("Merge gasRerpoterOutput.json files task", function() {
-  const projectPath = path.resolve(__dirname, "hardhat-merge-project");
+  const projectPath = path.resolve(__dirname, "../projects/hardhat-merge-project");
 
   useEnvironment(projectPath);
